detector/requirements.txt
flask==3.1.1
pyyaml==6.0.2
requests==2.32.3
psutil==7.0.0
gunicorn==23.0.0
